fix(decoderonly): load embed_tokens in the compile dtype
_create_embedding_layer builds an fp32 nn.Embedding and load_state_dict
keeps the module dtype, so a non-fp32 compiled model served its embedding
lookups in fp32. Every consumer had to patch over this at the call site
with .to(rbln_config.dtype), and paths without the patch (e.g. the runtime
wrapper's inputs_embeds preparation) fed fp32 embeddings to a runtime
compiled for bf16/fp16.

Cast embed_tokens to rbln_config.dtype right after loading, following the
existing gemma4 pattern, and drop the call-site casts that are now no-ops.
The cast in Exaone-4.5's forward stays: it also covers caller-passed
inputs_embeds.

+++ src/optimum/rbln/transformers/models/exaone4_5/modeling_exaone4_5.py
@@ -375,7 +375,7 @@ class RBLNExaone4_5_Model(RBLNDecoderOnlyModel):
         video_grid_thw: torch.LongTensor = None,
         second_per_grid_ts: torch.Tensor = None,
     ):
-        inputs_embeds = self.embed_tokens(input_ids).to(self.rbln_config.dtype)
+        inputs_embeds = self.embed_tokens(input_ids)
 
         if pixel_values is not None:
             image_embeds = self.visual(pixel_values, grid_thw=image_grid_thw)
